requirements.txt
certifi==2024.7.4
charset-normalizer==3.3.2
dateparser==1.2.0
facebook-sdk==3.1.0
idna==3.7
instagrapi==2.1.3
instaloader==4.13
numpy==1.26.4
pandas==2.1.4
pillow==10.4.0
pycryptodomex==3.20.0
pydantic==2.7.1
pydantic_core==2.18.2
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
regex==2024.11.6
requests==2.32.0
scrapetube==2.5.1
six==1.16.0
typing_extensions==4.12.2
tzdata==2024.1
tzlocal==5.2
urllib3==2.2.2
youtube-comment-downloader==0.1.76
twscrape~=0.15.0
openai~=1.61.0
demoji~=1.1.0
nltk~=3.9.1
matplotlib~=3.10.0
seaborn~=0.13.2
transformers~=4.48.3
wordcloud~=1.9.4